+++ controllers/controllersfakes/fake_registrar.go
@@ -5,9 +5,8 @@ import (
 	"context"
 	"sync"
 
-	"sigs.k8s.io/cluster-api-provider-gcp/api/v1beta1"
-
 	"github.com/giantswarm/dns-operator-gcp/controllers"
+	"sigs.k8s.io/cluster-api-provider-gcp/api/v1beta1"
 )
 
 type FakeRegistrar struct {